pyproject.toml
[tool.poetry]
name = "factory-bot"
version = "0.1.0"
description = "Telegram API + DRF bot"
authors = ["AlenGozy <=>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^4.2.5"
djangorestframework = "^3.14.0"
psycopg2-binary = "^2.9.7"
djangorestframework-simplejwt = "^5.3.0"
drf-spectacular = "^0.26.5"
python-telegram-bot = "13.3"
uuid = "^1.30"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
